% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leninature-package.R
\docType{package}
\name{leninature-package}
\alias{leninature}
\alias{leninature-package}
\title{leninature: Provides the works of Lenin in an analysis-ready format}
\description{
This package contains all of Lenin's works available on the Marxists Internet Archive (marxists.org) in the form of a tidy data frame.
}
\author{
\strong{Maintainer}: Circe Malone \email{circemalone@gmail.com}

}
\keyword{internal}
