% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_lenin_mentions.R
\name{count_lenin_mentions}
\alias{count_lenin_mentions}
\title{Count mentions of key words in Lenin's work}
\usage{
count_lenin_mentions(words, ignore_case = TRUE)
}
\arguments{
\item{words}{vector of key words to search for}

\item{ignore_case}{logical, should case sensitivity be considered?}
}
\value{
Returns the lenin data frame with two new columns added:
\describe{
\item{mentions}{number of occurrences of the key word in the \code{text} column}
\item{word}{key word}
}
The data frame will be in "long" format, with a new row for each key word.
}
\description{
Counts how often a key word occurs in Lenin's work. Considers word
boundaries, optionally considers letter case.
}
